A different implementation of a generic GLSL-compatible vector class, which provides three base classes (XY, XYZ, XYZW) and a vector class which is parametised on the base class to access its available members. Compared to the previous implementation, this is still a generic implementation but it moves the utility functions(abs(), etc) and operators inside the class definition. This makes the code shorter/simpler, and resolves some of the argument-dependant lookup (ADL) problems I was having with operators being defined outside the class.
However, it should be noted that this is intended to be a temporary implementation. I wanted to commit it as it is the most promising generic/templatised implementation so far, but actually I am expecting to move to three independent vector implementations (for 2, 3 and 4-element vectors). This will come in a future commit.